requirements.txt
streamlit
bcrypt
fpdf
python-dotenv
mysql-connector-python
openai
streamlit-chat
matplotlib
